embedded_systems/Project_Spotify_Controller/python/requirements.txt
pyserial
pyautogui
pycaw